Add movement physics tests and test helpers
- Add individual test files for forward movement, strafing left/right
- Update conftest.py with better test fixtures
- Enhance test_movement_physics.py with more comprehensive tests
- Add test_helpers.py for shared test utilities
- Add debug_action_format.py for debugging action formatting
- Remove obsolete test_visualize_movement.py

+++ tests/python/conftest.py
@@ -65,11 +65,13 @@ class RecordingWrapper:
         if self.step_count == 0:
             return None
             
-        # Stack all actions
+        # Stack all actions: shape is [steps, worlds, agents, 3]
         all_actions = np.stack(self.actions)
         
-        # Flatten in the order viewer expects
-        flat_actions = all_actions.flatten()
+        # The viewer expects data in order: [step][world][agent][3 components]
+        # numpy.flatten() uses C-order (row-major) by default, which is correct
+        # for this layout. Just ensure we're saving as int32.
+        flat_actions = all_actions.astype(np.int32).flatten()
         
         # Save binary file
         flat_actions.tofile(str(self.output_path))
